notif [nfc]: Move NotificationOpenPayload to a separate file
