+++ lib/notifications/open.dart
@@ -1,0 +1,85 @@
+import '../api/model/model.dart';
+import '../model/narrow.dart';
+
+/// The information contained in 'zulip://notification/…' internal
+/// Android intent data URL, used for notification-open flow.
+class NotificationOpenPayload {
+  final Uri realmUrl;
+  final int userId;
+  final Narrow narrow;
+
+  NotificationOpenPayload({
+    required this.realmUrl,
+    required this.userId,
+    required this.narrow,
+  });
+
+  factory NotificationOpenPayload.parseUrl(Uri url) {
+    if (url case Uri(
+      scheme: 'zulip',
+      host: 'notification',
+      queryParameters: {
+        'realm_url': var realmUrlStr,
+        'user_id': var userIdStr,
+        'narrow_type': var narrowType,
+        // In case of narrowType == 'topic':
+        // 'channel_id' and 'topic' handled below.
+
+        // In case of narrowType == 'dm':
+        // 'all_recipient_ids' handled below.
+      },
+    )) {
+      final realmUrl = Uri.parse(realmUrlStr);
+      final userId = int.parse(userIdStr, radix: 10);
+
+      final Narrow narrow;
+      switch (narrowType) {
+        case 'topic':
+          final channelIdStr = url.queryParameters['channel_id']!;
+          final channelId = int.parse(channelIdStr, radix: 10);
+          final topicStr = url.queryParameters['topic']!;
+          narrow = TopicNarrow(channelId, TopicName(topicStr));
+        case 'dm':
+          final allRecipientIdsStr = url.queryParameters['all_recipient_ids']!;
+          final allRecipientIds = allRecipientIdsStr.split(',')
+            .map((idStr) => int.parse(idStr, radix: 10))
+            .toList(growable: false);
+          narrow = DmNarrow(allRecipientIds: allRecipientIds, selfUserId: userId);
+        default:
+          throw const FormatException();
+      }
+
+      return NotificationOpenPayload(
+        realmUrl: realmUrl,
+        userId: userId,
+        narrow: narrow,
+      );
+    } else {
+      // TODO(dart): simplify after https://github.com/dart-lang/language/issues/2537
+      throw const FormatException();
+    }
+  }
+
+  Uri buildUrl() {
+    return Uri(
+      scheme: 'zulip',
+      host: 'notification',
+      queryParameters: <String, String>{
+        'realm_url': realmUrl.toString(),
+        'user_id': userId.toString(),
+        ...(switch (narrow) {
+          TopicNarrow(streamId: var channelId, :var topic) => {
+            'narrow_type': 'topic',
+            'channel_id': channelId.toString(),
+            'topic': topic.apiName,
+          },
+          DmNarrow(:var allRecipientIds) => {
+            'narrow_type': 'dm',
+            'all_recipient_ids': allRecipientIds.join(','),
+          },
+          _ => throw UnsupportedError('Found an unexpected Narrow of type ${narrow.runtimeType}.'),
+        })
+      },
+    );
+  }
+}
